Implemented JPEG file copy.
Using hard-coded settings that should be moved
in future to a settings file, the main loop now
opens each file, checks that it is a JPEG image,
checks its aspect ratio, and copies it into one of
three folders corresponding to the aspect ratio.

In the future, it will be desirable to have options
to include or exclude certain aspect ratios as well
as set minimum pixel counts. These settings
should also be ported to the settings file,
when it is implemented.